Re-organize general log items
