+++ crates/ergot/src/lib.rs
@@ -4,8 +4,8 @@
 
 pub mod address;
 pub mod book;
-pub mod fmtlog;
 pub mod interface_manager;
+pub mod logging;
 pub mod nash;
 pub mod net_stack;
 pub mod socket;
@@ -16,6 +16,9 @@ pub mod wire_frames;
 
 #[cfg(any(test, feature = "std"))]
 pub mod conformance;
+
+// Compat hack, remove on next breaking change
+pub use logging::fmtlog;
 
 pub use address::Address;
 use interface_manager::InterfaceSendError;
@@ -167,86 +170,6 @@ impl From<HeaderSeq> for Header {
 }
 
 pub const DEFAULT_TTL: u8 = 16;
-
-use logger::logger;
-pub use logger::set_logger_racy;
-
-mod logger {
-    use core::sync::atomic::Ordering;
-
-    use log::Log;
-    use portable_atomic::AtomicUsize;
-
-    static NOP_LOGGER: NopLogger = NopLogger;
-    static mut LOGGER: StaticLogger = StaticLogger::new();
-
-    #[allow(static_mut_refs)]
-    pub unsafe fn set_logger_racy(logger: &'static dyn Log) -> Result<(), ()> {
-        unsafe { LOGGER.set_logger_racy(logger) }
-    }
-
-    #[allow(static_mut_refs)]
-    pub fn logger() -> &'static dyn Log {
-        unsafe { &mut LOGGER }.logger()
-    }
-
-    struct StaticLogger {
-        logger: &'static dyn Log,
-        state: AtomicUsize,
-    }
-
-    impl StaticLogger {
-        pub const fn new() -> Self {
-            Self {
-                logger: &NOP_LOGGER,
-                state: AtomicUsize::new(State::Uninitialized as usize),
-            }
-        }
-
-        pub unsafe fn set_logger_racy(&mut self, logger: &'static dyn Log) -> Result<(), ()> {
-            match unsafe {
-                core::mem::transmute::<usize, State>(self.state.load(Ordering::Acquire))
-            } {
-                State::Initialized => Err(()),
-                State::Uninitialized => {
-                    self.logger = logger;
-                    self.state
-                        .store(State::Initialized as usize, Ordering::Release);
-                    Ok(())
-                }
-                State::Initializing => {
-                    unreachable!("set_logger_racy should not be used with other logging functions")
-                }
-            }
-        }
-
-        pub fn logger(&self) -> &'static dyn Log {
-            if self.state.load(Ordering::Acquire) != State::Initialized as usize {
-                &NOP_LOGGER
-            } else {
-                self.logger
-            }
-        }
-    }
-
-    #[repr(usize)]
-    enum State {
-        Uninitialized,
-        Initializing,
-        Initialized,
-    }
-
-    pub struct NopLogger;
-
-    impl Log for NopLogger {
-        fn enabled(&self, _: &log::Metadata) -> bool {
-            false
-        }
-
-        fn log(&self, _: &log::Record) {}
-        fn flush(&self) {}
-    }
-}
 
 /// Exports of used crate versions
 pub mod exports {
